pyproject.toml
[build-system]
requires = [
    "maturin>=1.6,<2.0",
]
build-backend = "maturin"

[[project.authors]]
name = "hab"
email = "_@habnab.it"

[project.urls]
Homepage = "https://github.com/habnabit/passacre"
Repository = "https://github.com/habnabit/passacre"

[project]
name = "passacre"
description = "better repeatable password generation"
readme = "README.rst"
license = { file = "COPYING" }

# supporting 2.7 (again?) means an older maturin
# requires-python = ">=2.7"
requires-python = ">=3.7"
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: ISC License (ISCL)",
    "Operating System :: OS Independent",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dynamic = ["version"]

[project.optional-dependencies]
yaml = ["PyYAML"]
clipboard = ["xerox"]
keccak = []
skein = []
yubikey = ["ykpers-cffi"]
all = ["PyYAML", "xerox", "ykpers-cffi"]

[project.scripts]
passacre = "passacre.application:main"

[tool.maturin]
features = ["pyo3/extension-module"]
python-source = "python"
module-name = "passacre._pyo3_backend"
